refactor: move the lease file helpers into one `lease` module
